+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file. This projec
 generic adapter.
 - Extracted the logic for an Eloquent `hasManyThrough` relation into its own relationship adapter (was
 previously in the `has-many` adapter).
+- Moved the `FindsManyResources` trait from the `Store` namespace to `Adapter\Concerns`.
 
 ## [1.0.0-alpha.2] - 2018-05-06